langchain_QA_ChatBot/requirements.txt
# LangChain
langchain

# Google Gemini (PaLM) Integration
langchain-google-genai

# Environment Variable Management
python-dotenv
streamlit